refactor: decompose infrastructure/services.py into facade package (#2050)
